:sparkles: Add change password modal
